pipeline/pyproject.toml
[tool.poetry]
name = "checkstatusgovbr-pipeline"
version = "0.1.0"
description = "ETL Pipeline of project Checkstatus .gov.br"
authors = ["Gilson Filho <me@gilsondev.in>"]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
pandas = "^1.5.0"
psycopg2-binary = "^2.9.4"
whoisit = "^2.6.5"
tenacity = "^8.1.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
mypy = "^0.982"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"

[tool.poetry.group.dev.dependencies]
responses = "^0.22.0"
types-psycopg2 = "^2.9.21.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
